<Refactor>[Pyut]: <Use modern properties>
[https://github.com/hasii2011/PyUt/issues/305]

+++ src/org/pyut/ogl/sd/OglSDMessage.py
@@ -152,8 +152,8 @@ class OglSDMessage(OglLink):
         Returns:  A tuple of anchor points for the source and destination shapes
 
         """
-        srcY: float = pyutSDMessage.getSrcY() - srcShape.getLifeLineShape().GetPosition()[1]
-        dstY: float = pyutSDMessage.getDstY() - dstShape.getLifeLineShape().GetPosition()[1]
+        srcY: int = pyutSDMessage.getSrcY() - srcShape.getLifeLineShape().GetPosition()[1]
+        dstY: int = pyutSDMessage.getDstY() - dstShape.getLifeLineShape().GetPosition()[1]
 
         srcAnchor: AnchorPoint = srcShape.getLifeLineShape().AddAnchor(0, srcY)
         dstAnchor: AnchorPoint = dstShape.getLifeLineShape().AddAnchor(0, dstY)
